hardware. nafarr: memory: hyperbus: Fix Controller Bug
Use the 'fire' signal instead for 'ready' only to prevent the
controller locks up with instruction and data requests.

The controller was empting the AXI ID FIFO and would return the
last element. This would return instruction or data responses
with the wrong ID and, for example, return instructions to the
data port.

Signed-off-by: Daniel Schultz <dnltz@aesc-silicon.de>