hardware: nafarr: Fix WishboneConfig
Force the WishboneConfig for each peripheral to use the 10 or 20
address bits, depending whether the IP is using 4KB or 4MB address
space.

The SpinalHDL Wishbone configuration is removing the two LSB of the
address bus. Thus, each peripheral should also remove two bits.
Otherwise, the address from the WishboneDecoder might be passed to
the peripheral mapper and registers won't match anymore.

Signed-off-by: Daniel Schultz <dnltz@aesc-silicon.de>